+++ R/colony-growth.R
@@ -7,20 +7,16 @@
 #' @param taus an optional vector of taus to test. If not supplied, `seq(min(t),
 #'   max(t), length.out = 50)` will be used
 #' @param t the unquoted column name for the time variable in `data`
-#' @param formula a formula passed to `glm`.  This should include the time
+#' @param formula a formula passed to `glm()` or `glm.nb()`.  This should include the time
 #'   variable supplied to `t`
-#' @param family the model family to use.  By default, the data are fit with a
-#'   log-link gaussian generalized linear model. Because a log link is used, the
-#'   response variable should not be log-transformed.  For count data (e.g.
-#'   number of workers), use "poisson".  For overdispersed count data, use
-#'   "negbin" to fit the model using `glm.nb()` from the `MASS` package.
+#' @param family passed to `glm()`
+#' @param ... additional arguments passed to `glm()` or `glm.nb()`
 #' @return a tibble with a column for the winning tau and a column for the
 #'   winning model
 #'
 #' @import dplyr
 #' @import rlang
 #' @importFrom stats update logLik terms glm poisson as.formula gaussian
-#' @importFrom MASS glm.nb
 #'
 #' @keywords internal
 #'
@@ -32,12 +28,13 @@
 #' # Using weeks
 #' brkpt(testbees, t = week, formula = mass ~ week)
 #' }
-brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson", "negbin")) {
+brkpt <- function(data, taus = NULL, t, formula, family = gaussian(link = "log"), ...) {
   #TODO: make sure none of the variables are called '.post'?
   fterms <- terms(formula)
   t <- enquo(t)
   tvar <-as_name(t)
-  fam <- match.arg(family)
+  # fam <- enquo(family)
+  more_args <- list2(...)
 
   if (is.null(taus)) {
     tvec <- data[[tvar]]
@@ -68,43 +65,13 @@ brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson
   # adds `.post` to formula. Would not be difficult to modify for other interactions
   f <- update(formula, ~. + .post)
   LLs <- c()
-  if (fam == "gaussian") {
-    for (i in 1:length(taus)) {
-      usetau <- taus[i]
-      data2 <- mutate(data, .post = ifelse(!!t <= usetau, 0, !!t - usetau))
 
-      m0 <- try(glm(f, family = gaussian(link = "log"), data = data2))
-      if (!inherits(m0, "try-error")) {
-        LLs[i] <- logLik(m0)
-      } #else?
-      #TODO: what if there is an error?
-      # LLs
-    }
-  } else if (fam == "poisson") {
-    for (i in 1:length(taus)) {
-      usetau <- taus[i]
-      data2 <- mutate(data, .post = ifelse(!!t <= usetau, 0, !!t - usetau))
 
-      m0 <- try(glm(f, family = "poisson", data = data2))
-      if (!inherits(m0, "try-error")) {
-        LLs[i] <- logLik(m0)
-      } #else?
-      #TODO: what if there is an error?
-      # LLs
-    }
-
-  } else if (fam == "negbin") {
-    for (i in 1:length(taus)) {
-      usetau = taus[i]
-      data2 <- mutate(data, .post = ifelse(!!t <= usetau, 0, !!t - usetau))
-
-      m0 = try(MASS::glm.nb(f, data = data2))
-      if (!inherits(m0, "try-error")) {
-        LLs[i] = logLik(m0)
-      } #else?
-      #TODO: what if there is an error?
-      # LLs
-    }
+  for (i in 1:length(taus)) {
+    usetau <- taus[i]
+    data2 <- mutate(data, .post = ifelse(!!t <= usetau, 0, !!t - usetau))
+    m0 <- exec("glm", formula = f, family = family, data = data2, !!!more_args)
+    LLs[i] <- logLik(m0)
   }
 
   tau_win <- taus[which(LLs == max(LLs))]
@@ -116,13 +83,76 @@ brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson
   #TODO: I don't really like that it re-fits the model.  I could have it save them all and only re-fit in the case of a tau tie.
   data_win <- mutate(data, .post = ifelse(!!t <= tau_win, 0, !!t - tau_win))
 
-  if (fam == "gaussian") {
-    m_win <- glm(f, family = gaussian(link = "log"), data = data_win)
-  } else if (fam == "poisson") {
-    m_win <- glm(f, family = poisson(link = "log"), data = data_win)
-  } else if (fam == "negbin") {
-    m_win <- glm.nb(f, data = data_win)
+  m_win <- exec("glm", formula = f, family = family, data = data_win, !!!more_args)
+
+  return(tibble(tau = tau_win, model = list(m_win)))
+}
+
+
+#' @describeIn brkpt
+#' @param link passed to `glm.nb()`
+#'
+#' @import dplyr
+#' @import rlang
+#' @importFrom stats update logLik terms as.formula
+#' @importFrom MASS glm.nb
+#'
+#' @keywords internal
+#'
+brkpt.nb <- function(data, taus = NULL, t, formula, link = "log", ...) {
+  #TODO: make sure none of the variables are called '.post'?
+  fterms <- terms(formula)
+  t <- enquo(t)
+  tvar <-as_name(t)
+  # fam <- enquo(family)
+  more_args <- list2(...)
+
+  if (is.null(taus)) {
+    tvec <- data[[tvar]]
+    taus <- seq(min(tvec), max(tvec), length.out = 50)
   }
+
+  #Check that time variable is in the formula
+  if (!tvar %in% attr(fterms, "term.labels")) {
+    abort(paste0("'",tvar,"' is missing from the model formula"))
+  }
+
+  #Check that at least some taus are in range of t
+  if (all(taus > max(data[[tvar]])) | all(taus < min(data[[tvar]]))) {
+    abort(paste0("At least one tau must be in range of '", tvar, "'"))
+  }
+  #If some taus are out of range of t, drop them
+  if (any(taus > max(data[[tvar]])) |
+      any(taus < min(data[[tvar]]))) {
+    warning(paste0(
+      "Some taus were not used because they were outside of range of '",
+      tvar,
+      "'"
+    ))
+    taus <-
+      taus[taus <= max(data[[tvar]]) & taus >= min(data[[tvar]])]
+  }
+
+  # adds `.post` to formula. Would not be difficult to modify for other interactions
+  f <- update(formula, ~. + .post)
+  LLs <- c()
+
+    for (i in 1:length(taus)) {
+      usetau <- taus[i]
+      data2 <- mutate(data, .post = ifelse(!!t <= usetau, 0, !!t - usetau))
+      m0 <- exec("glm.nb", formula = f, data = data2, link = link, !!!more_args)
+      LLs[i] <- logLik(m0)
+    }
+
+  tau_win <- taus[which(LLs == max(LLs))]
+
+  # if multiple equivalent taus are found, this should fail
+  if (length(tau_win) > 1) {
+    abort("More than one equivalent tau found")
+  }
+  #TODO: I don't really like that it re-fits the model.  I could have it save them all and only re-fit in the case of a tau tie.
+  data_win <- mutate(data, .post = ifelse(!!t <= tau_win, 0, !!t - tau_win))
+  m_win <- exec("glm.nb",formula = f, data = data_win, !!!more_args)
   return(tibble(tau = tau_win, model = list(m_win)))
 }
 
@@ -144,14 +174,11 @@ brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson
 #'   response is your measure of colony growth, time is whatever measure of time
 #'   you have (date, number of weeks, etc.) and covariates are any optional
 #'   co-variates you want to fit at the colony level.
-#' @param family the model family to use.  By default, the data are fit with a
-#'   log-link gaussian generalized linear model. Because a log link is used, the
-#'   response variable should not be log-transformed.  For count data (e.g.
-#'   number of workers), use "poisson".  For overdispersed count data, use
-#'   "negbin" to fit the model using `glm.nb()` from the `MASS` package.
+#' @param family passed to `glm()`.
 #' @param augment when FALSE, `bumbl` returns a summary dataframe with one row
 #'   for each colonyID.  When TRUE, it returns the original data with additional
 #'   columns containing model coefficients.
+#' @param ... additional arguments passed to `glm()` or `glm.nb()`.
 #'
 #' @details Colony growth is modeled as increasing exponentialy until the colony
 #'   switches to gyne production, at which time the workers die and gynes leave
@@ -185,7 +212,6 @@ brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson
 #' @importFrom purrr map map_dbl map2_df
 #' @import broom
 #' @importFrom glue glue
-#' @importFrom utils packageVersion
 #' @export
 #'
 #' @examples
@@ -196,13 +222,14 @@ brkpt <- function(data, taus = NULL, t, formula, family = c("gaussian", "poisson
 #'
 #' bombus2 <- bombus[bombus$colony != 67, ]
 #' bumbl(bombus2, colonyID = colony, t = week, formula = mass ~ week)
-bumbl <- function(data, colonyID = NULL, t, formula, family = c("gaussian", "poisson", "negbin"), augment = FALSE, taus = NULL) {
+bumbl <- function(data, colonyID = NULL, t, formula, family = gaussian(link = "log"), augment = FALSE, taus = NULL, ...) {
   #TODO: scoop up all the warnings from brkpt() and present a summary at the
   #end.
 
   colonyID <- enquo(colonyID)
   t <- enquo(t)
-  fam <- match.arg(family)
+  # family <- enquo(family)
+  more_args <- list2(...)
 
   if (quo_is_null(colonyID)) {
     df <- data
@@ -239,10 +266,129 @@ bumbl <- function(data, colonyID = NULL, t, formula, family = c("gaussian", "poi
   resultdf <-
     purrr::map2_df(dflist,
                    names(dflist),
-                   ~brkpt_w_err(brkpt(.x, taus = {{taus}},
+                   ~brkpt_w_err(brkpt(data = .x,
+                                      taus = {{taus}},
                                       t = !!t,
                                       formula = formula,
-                                      family = fam),
+                                      family = family,
+                                      !!!more_args),
+                                .y),
+                   .id = as_name(colonyID))
+
+  predictdf <-
+    resultdf %>%
+    dplyr::select(-"tau") %>%
+    mutate(aug = purrr::map(.data$model, broom::augment)) %>%
+    unnest(.data$aug) %>%
+    dplyr::select(!!colonyID, !!t, ".fitted", ".se.fit", ".resid")
+
+  modeldf <-
+    resultdf %>%
+    mutate(coefs = purrr::map(.data$model, broom::tidy)) %>%
+    unnest(.data$coefs) %>%
+    dplyr::select(!!colonyID, "tau", "model", "term", "estimate") %>%
+    spread(key = "term", value = "estimate") %>%
+    mutate(logNmax = purrr::map_dbl(.data$model, ~ max(predict(.), na.rm = TRUE))) %>%
+    dplyr::select(-"model") %>%
+    dplyr::select(
+      !!colonyID,
+      "tau",
+      logN0 = "(Intercept)",
+      logLam = !!t,
+      decay = ".post",
+      "logNmax",
+      everything()
+    )
+
+  if (augment == TRUE) {
+    augmented_df <-
+      left_join(df, modeldf, by = as_name(colonyID))
+
+    full_augmented_df <-
+      left_join(augmented_df, predictdf, by = c(as_name(colonyID), as_name(t)))
+
+    #add attributes
+    attr(full_augmented_df, "colonyID") <- as_name(colonyID)
+    attr(full_augmented_df, "t") <- as_name(t)
+    attr(full_augmented_df, "formula") <- formula
+    attributes(full_augmented_df)
+
+    class(full_augmented_df) <- c(class(full_augmented_df), "bumbldf")
+
+    return(full_augmented_df)
+
+  } else {
+    return(modeldf)
+  }
+
+}
+
+
+
+
+
+#' @describeIn bumbl
+#'
+#' @param link passed to `glm.nb()` from the `MASS` package
+#'
+#' @import tidyr
+#' @import rlang
+#' @import dplyr
+#' @importFrom purrr map map_dbl map2_df
+#' @import broom
+#' @importFrom glue glue
+#' @export
+#'
+bumbl.nb <- function(data, colonyID = NULL, t, formula, link = "log", augment = FALSE, taus = NULL, ...) {
+  #TODO: scoop up all the warnings from brkpt() and present a summary at the
+  #end.
+
+  colonyID <- enquo(colonyID)
+  t <- enquo(t)
+  # family <- enquo(family)
+  more_args <- list2(...)
+
+  if (quo_is_null(colonyID)) {
+    df <- data
+    dflist <- list("NA" = data)
+    colonyID <- "colony"
+
+  } else {
+    df <-
+      data %>%
+      # make sure colonyID is a character vector
+      mutate(!!colonyID := as.character(!!colonyID)) %>%
+      group_by(!!colonyID)
+
+    dflist <- group_split(df)
+    names(dflist) <- group_keys(group_by(data, !!colonyID))[[1]]
+  }
+
+  model_list <- vector("list", length(dflist))
+  names(model_list) <- names(dflist)
+
+  brkpt_w_err <- function(code, colonyID) {
+    tryCatch(
+      code,
+      error = function(c) {
+        message(
+          glue::glue(
+            "Warning: {c$message} for colonyID '{colonyID}'. Omitting from results."
+          )
+        )
+      }
+    )
+  }
+
+  resultdf <-
+    purrr::map2_df(dflist,
+                   names(dflist),
+                   ~brkpt_w_err(brkpt.nb(data = .x,
+                                      taus = {{taus}},
+                                      t = !!t,
+                                      formula = formula,
+                                      link = link,
+                                      !!!more_args),
                                 .y),
                    .id = as_name(colonyID))