+++ NEWS.md
@@ -1,8 +1,11 @@
 # bumbl (development version)
-* changed `bumbl()` to use a log-link glm by default instead of an lm.  Use untransformed response variables.
-* added argument to `bumbl()` to allow for count data
+* changed `bumbl()` to use `glm()` instead of `lm()` to fit models.  Uses log-link gaussian error by default.
+* `family` argument in `bumbl()` now passed directly to `glm()`
+* Added `bumbl.nb()`  for overdispersed count data.  It uses `MASS::glm.nb()` to fit models.
+* Added `...` argument to `bumbl()` and `bumbl.nb()` to pass additional arguments (e.g. offset) to `glm()` and `glm.nb()` respectively
 * added change in mass to `bombus` data
 * now requires tidyr version 1.0.0 or greater
+
 
 # bumbl 0.0.1
 * `bumbl()` no longer errors when a single colony produces an error